requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
jinja2
python-multipart
python-dotenv
email-validator
pytest